+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Make some setting dialogs resizable.
 - Allow custom syntaxes to be deleted by dragging them to the Trash.
 - Allow CotEditor Syntax files and multiple replacement definitions to be imported by dropping them onto their lists.
+- Add the context menu to the list in Mode settings pane.
 - [dev] Tweak the accent color.
 - [dev] Update the build environment to Xcode 27 Beta (Swift 6.4, macOS 27 SDK).